added auth to journal endpoints

+++ src/main/java/com/manan565/journalApp/services/JournalEntryService.java
@@ -5,7 +5,6 @@ import com.manan565.journalApp.entity.User;
 import com.manan565.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.annotation.Id;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -27,14 +26,12 @@ public class JournalEntryService {
         journalEntry.setDate(LocalDateTime.now());
         JournalEntry saved= journalEntryRepository.save(journalEntry);
         user.getJournalEntries().add(saved);
-        userService.saveEntry(user);
+        userService.saveUser(user);
     }
     public void saveEntry( JournalEntry journalEntry){
 
         JournalEntry saved= journalEntryRepository.save(journalEntry);
     }
-
-
 
 
     public List<JournalEntry> getAll(){
@@ -45,10 +42,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id,String userName){
-         User user =userService.findByUserName((userName));
-         user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-         userService.saveEntry(user);
-         journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id,String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName((userName));
+            removed=user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleteing the entry.", e);
+        }
+
+        return removed;
     }
 }
